pyproject.toml
[project]
name = "aq3d-bots"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.6"
dependencies = [
    "numpy==2.2.6",
    "opencv-python==4.12.0.88",
    "pillow==12.0.0",
    "pynput==1.8.1",
    "pywin32==311",
]
